@@ -7,7 +7,8 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
+  before_action :warn_about_email_whitelist
 
   private
 
@@ -17,10 +18,9 @@ class ApplicationController < ActionController::Base
 
       # In the development environment, your current_user will be the
       # first User in the database or a dummy one created below.
-      if !Rails.env.production?
+      if Rails.env.development? || Rails.env.test?
         user = User.first_or_create(email: "alvar@hanso.dk", name: "Alvar Hanso")
         session[:user_id] = user.id
-        flash[:notice] = "Signed in!"
       else
         user = User.find(session[:user_id]) if session[:user_id].present?
       end
@@ -39,14 +39,15 @@ class ApplicationController < ActionController::Base
   helper_method :user_signed_in?
 
   def authenticate_user!
-    if current_user
-      true
-    else
-      session["return_to"] ||= request.url
-      redirect_to login_path unless login_redirect? or oauth_callback?
-    end
+    return true if current_user
+    session["return_to"] ||= request.url
+    redirect_to login_path unless redirect_loop?
   end
   helper_method :authenticate_user!
+
+  def redirect_loop?
+    login_redirect? || oauth_callback?
+  end
 
   def login_redirect?
     request.path == login_path
@@ -54,5 +55,11 @@ class ApplicationController < ActionController::Base
 
   def oauth_callback?
     request.path == oauth_callback_path("google_oauth2")
+  end
+
+  def warn_about_email_whitelist
+    if Rails.env.production? && !User.email_whitelist_enabled?
+      flash[:error] = "WARNING: email whitelisting is currently disabled, set ENV['ORIENTATION_EMAIL_WHITELIST'] to enable it."
+    end
   end
 end
