@@ -1,4 +1,4 @@
-class ArticleSubscription < ActiveRecord::Base
+class ArticleSubscription < ApplicationRecord
   include Dateable
 
   belongs_to :article, counter_cache: :subscriptions_count
@@ -6,20 +6,20 @@ class ArticleSubscription < ActiveRecord::Base
 
   def send_update
     clear_existing_queued_update_jobs if update_queued?
-    Delayed::Job.enqueue(SendArticleUpdateJob.new(self.article_id, user.id), run_at: 5.minutes.from_now)
+    SendArticleUpdateJob.set(wait: 5.minutes).perform_later(article.id, user.id)
   end
 
   def update_queued?
-    existing_queued_update_jobs.size.nonzero? ? true : false
+    existing_queued_update_jobs.any?
   end
 
   def existing_queued_update_jobs
-    matching_jobs = Delayed::Job.select do |job|
-      job.handler =~ /SendArticleUpdateJob/ && job.handler =~ /article_id: #{self.article_id}/
-    end
+    Delayed::Job
+      .where(%Q["delayed_jobs"."handler" LIKE '%SendArticleUpdateJob%'])
+      .where(%Q["delayed_jobs"."handler" LIKE '%arguments:%#{article_id}%'])
   end
 
   def clear_existing_queued_update_jobs
-    existing_queued_update_jobs.each(&:destroy)
+    existing_queued_update_jobs.destroy_all
   end
 end
