@@ -2,59 +2,67 @@ class ArticleDecorator < ApplicationDecorator
   delegate_all
 
   def authors
-    output = "#{author.link_tag} #{created_at_tag} "
+    output = [author.link_tag, created_at_tag]
 
     if editor && updated_at_tag
-      if editor.id != source.author_id
-        output << "and updated by #{editor_email_tag} on #{updated_at_tag}"
-      else
-        output << "and updated on #{updated_at_tag}"
-      end
+      output << "and updated"
+      output << "by #{editor_email_tag}" if editor.id != object.author_id
+      output << "on #{updated_at_tag}"
     end
 
-    output.html_safe
+    output.join(" ").html_safe
   end
 
   def title
-    source.try(:title) || "No title"
+    object.try(:title) || "No title"
   end
 
   def created_at_tag
-    time_element(source.created_at)
+    time_element(object.created_at)
   end
 
   def updated_at_tag
-    time_element(source.updated_at)
+    time_element(object.updated_at)
   end
 
   def author
-    AuthorDecorator.new(source.author)
+    AuthorDecorator.new(object.author)
   end
 
   def editor
-    AuthorDecorator.new(source.editor) if source.editor
+    AuthorDecorator.new(object.editor) if object.editor
   end
 
   def editor_email_tag
-    mail_to source.try(:editor).try(:email), source.try(:editor).try(:name)
+    mail_to object.try(:editor).try(:email), object.try(:editor).try(:name)
   end
 
   def freshness
-    content_tag(:span, "fresh", class: "state fresh") if source.fresh?
+    content_tag(:span, "fresh", class: "state fresh") if object.fresh?
   end
 
   def staleness
-    content_tag(:span, "stale", class: "state stale") if source.stale?
+    content_tag(:span, "stale", class: "state stale") if object.stale?
   end
 
   def rottenness
-    content_tag(:span, "rotten", class: "state rotten") if source.rotten?
+    content_tag(:span, "rotten", class: "state rotten") if object.rotten?
+  end
+
+  def rot_reporter
+    if object.rot_reporter
+      link_to AuthorDecorator.new(object.rot_reporter), author_url(object.author)
+    end
+  end
+
+  def rotted_at
+    object.rotted_at.to_date.to_s(:long_ordinal)
   end
 
   def signal
-    state = 'fresh' if source.fresh?
-    state = 'stale' if source.stale?
-    state = 'rotten' if source.rotten?
+    state = 'fresh' if object.fresh?
+    state = 'stale' if object.stale?
+    state = 'rotten' if object.rotten?
 
     screen_reader_text = content_tag(:span, state, class: 'srt')
 
