@@ -2,10 +2,10 @@ class ArticleSubscriptionDecorator < ApplicationDecorator
   delegate_all
 
   def created_at
-    source.created_at.to_formatted_s(:long)
+    object.created_at.to_formatted_s(:long)
   end
 
   def user
-    source.user.decorate
+    object.user.decorate
   end
 end
