@@ -2,7 +2,7 @@ class UserDecorator < ApplicationDecorator
   delegate_all
 
   def email
-    source.try :email || "No Email"
+    object.try :email || "No Email"
   end
 
   def email_tag
@@ -10,30 +10,30 @@ class UserDecorator < ApplicationDecorator
   end
 
   def first_name
-    source.name.split(" ").first
+    object.name.split(" ").first
   end
 
   def image
-    source.try(:image) or asset_path("default_avatar.jpg")
+    object.try(:image) or asset_path("default_avatar.jpg")
   end
 
   def image_link(size = 40)
-    link_to image_tag(image, class: 'thumb dib', height: size, width: size), author_path(source)
+    link_to image_tag(image, class: 'thumb dib', height: size, width: size), author_path(object)
   end
 
   def link
-    author_path(source)
+    author_path(object)
   end
 
   def link_tag
-    link_to name, author_path(source)
+    link_to name, author_path(object)
   end
 
   def name
-    source.try :name || "Anonymous"
+    object.try :name || "Anonymous"
   end
 
   def status
-    if source.active then "active" else "inactive" end
+    if object.active then "active" else "inactive" end
   end
 end
