@@ -1,9 +1,11 @@
-class SendArticleEndorsementJob < Struct.new(:endorsement_id)
-  def perform
-    endorsement = ArticleEndorsement.find(endorsement_id)
-    contributors = endorsement.article.contributors
-    article = endorsement.article
-    endorser = endorsement.user
+class SendArticleEndorsementJob < ApplicationJob
+  queue_as :default
+
+  def perform(endorsement_id)
+    endorsement  = ArticleEndorsement.find(endorsement_id)
+    article      = endorsement.article
+    endorser     = endorsement.user
+    contributors = article.contributors
 
     ArticleMailer.send_endorsement_notification_for(article, contributors, endorser).deliver
   end
