@@ -6,6 +6,7 @@ require "action_controller/railtie"
 require "action_mailer/railtie"
 require "sprockets/railtie"
 # require "rails/test_unit/railtie"
+require "action_cable/engine"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -17,34 +18,16 @@ module Orientation
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
-    # Custom directories with classes and modules you want to be autoloadable.
-    # config.autoload_paths += %W(#{config.root}/extras)
-
-    # Configure sensitive parameters which will be filtered from the log file.
-    config.filter_parameters += [:password]
-
     # Use SQL instead of Active Record's schema dumper when creating the database.
     # This is necessary if your schema can't be completely dumped by the schema dumper,
     # like if you have constraints or database-specific column types.
     config.active_record.schema_format = :sql
 
-    # Enable the asset pipeline.
-    config.assets.enabled = true
-
-    # Version of your assets, change this if you want to expire all your assets.
-    config.assets.version = '1.0'
-
-    # Precompilation assets list
-    config.assets.precompile += %w( vendor.js )
-
-    # Asset paths list
-    config.assets.paths << Rails.root.join('bower_components')
-
-    # Prevent Rails from initializing during the precompile phase on Heroku
-    config.assets.initialize_on_precompile = false
-
     # prefer indented sass syntax
     config.sass.preferred_syntax = :sass
+
+    # https://robots.thoughtbot.com/content-compression-with-rack-deflater
+    config.middleware.use Rack::Deflater
 
     config.generators do |g|
       g.test_framework :rspec
@@ -52,10 +35,18 @@ module Orientation
       g.helper_specs false
     end
 
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    config.active_job.queue_adapter = :delayed_job
 
-    # https://robots.thoughtbot.com/content-compression-with-rack-deflater
-    config.middleware.use Rack::Deflater
+    # We load environment-specific configuration values from 
+    # config/orientation.yml into Rails.configuration.orientation
+    #
+    # You can find example values in config/orientation.example.yml
+    #
+    # Once the application has booted you can assign values directly like this:
+    #   Rails.configuration.orientation["mailers"] = :mandrill
+    # And you can read those same values with:
+    #   Rails.configuration.orientation["mailers"]
+    #
+    config.orientation = config_for(:orientation)
   end
 end
