@@ -1,7 +1,7 @@
-if ENV["SLACK_WEBHOOK_URL"]
-  $slack = Slack::Notifier.new ENV["SLACK_WEBHOOK_URL"],
+SLACK_CONFIG = if ENV["SLACK_WEBHOOK_URL"]
+  Slack::Notifier.new ENV["SLACK_WEBHOOK_URL"],
   username: "Orientation",
   icon_emoji: ":book:"
 else
-  $slack = nil
+  nil
 end
