@@ -22,15 +22,21 @@ RSpec.describe ArticleMailer do
     it { is_expected.to be_from(email: 'orientation@codeschool.com') }
   end
 
-  context ".send_updates_for(article, user)" do
+  context ".send_updates_for(article, user, editor)" do
     let(:article) { create(:article) }
+    let(:other_user) { create(:user) }
     let(:mailer) { described_class.send_updates_for(article, user) }
+
+    before do
+      article.editor = other_user
+      article.reload
+    end
 
     subject { mailer }
 
     it { is_expected.to send_email_to(email: user.email) }
     it { is_expected.to use_template('article-subscription-update') }
-    it { is_expected.to have_subject("#{article.title} was just updated") }
+    it { is_expected.to have_subject("#{article.title} was updated by #{article.editor}") }
     it { is_expected.to be_from(email: 'orientation@codeschool.com') }
   end
 
@@ -41,14 +47,15 @@ RSpec.describe ArticleMailer do
         { name: user.name, email: user.email }
       end
     end
+    let(:reporter) { create(:user) }
 
-    let(:mailer) { described_class.send_rotten_notification_for(article, contributors) }
+    let(:mailer) { described_class.send_rotten_notification_for(article, contributors, reporter) }
 
     subject { mailer }
 
     it { is_expected.to send_email_to(email: contributors.first[:email]) }
     it { is_expected.to use_template('article-rotten-update') }
-    it { is_expected.to have_subject('Article Rotten Update') }
+    it { is_expected.to have_subject("#{reporter.name} marked #{article.title} as rotten") }
     it { is_expected.to be_from(email: 'orientation@codeschool.com') }
   end
 
